Syntax analyzer
Beginning work on the next portion of the compiler which is the syntax analyzer. By using the symbol table , and token array created from the lexical analyzer, we can begin putting elements/tokens of the code into a pushdown stack.

Still to be implemented: operator precedence table, and logic to know when to pop from stack, as well as the resulting switch:case which will hold machine language code templates.